Tell c_read_binary that it's decoding characters
These type and size hints let cython optimize the string decoding operation
to a pair of `__Pyx_PyObject_AsString` / `__Pyx_decode_c_string()` calls. This
is a lot more efficient than the previous code generation which invoked the
"encode" method on the abstract "py_data" object.

The performance improvement here varies based on how much string data exists
in the input buffer, but the benchmark suite shows a 40%+ speed improvement
with this change.