Recognize the `cpp_include` header keyword
`cpp_include` adds a custom C++ include to the output of the C++ code
generator for this Thrift document. While (obviously) not used by the
Python implementation, it can still appear in Thrift IDL files that we
need to parse, so we recognize the keyword and just discard the value.

See also: https://thrift.apache.org/docs/idl#c-include